5.1 Отрисуй меня полностью (часть 1) с изменениями
